feat: handleClick 함수를 통한 api 연결이 버튼 내부가 아닌 상위 컴포넌트에서 일어나도록 수정
